Moves ExperimentCard component to experiments folder
